fix: pin mitt to 2.1.0 (CommonJS) and restore preload.js
