+++ __tests__/renderer.init.test.js
@@ -8,7 +8,7 @@ jest.mock('../chartWorker.js', () => ({
 }));
 
 test('renderer bootstraps without errors', async () => {
-  await import('../src/preload.mjs');
+  await import('../src/preload.js');
   global.window = { api: contextBridge.exposeInMainWorld.mock.calls[0][1], electronAPI: { getVersion: jest.fn(() => Promise.resolve('1.0.0')) } };
   const bus = global.window.api.bus;
   expect(bus).toBeDefined();
