+++ package.json
@@ -18,7 +18,7 @@
     "jsdom": "^26.1.0"
   },
   "dependencies": {
-    "mitt": "^3.0.1",
+    "mitt": "2.1.0",
     "papaparse": "^5.5.3"
   },
   "build": {
